fix: 게시글 상세 페이지에도 베너 추가
