+++ src/app/posts/page.tsx
@@ -16,30 +16,6 @@ import Tab from '@/shared/ui/tab';
 import { BellButton, SiteHeader } from '@/widgets/header';
 import { useCallback, useState } from 'react';
 
-const TODAY_CONSULTANT_BANNERS = [
-  {
-    id: 1,
-    rank: 1,
-    name: '김혜정',
-    avatarUrl: 'https://picsum.photos/200/300',
-    shop: '아이디헤어문정점',
-  },
-  {
-    id: 2,
-    rank: 2,
-    name: '김혜정2',
-    avatarUrl: 'https://picsum.photos/200/300',
-    shop: '아이디헤어문정점2',
-  },
-  {
-    id: 3,
-    rank: 3,
-    name: '김혜정2',
-    avatarUrl: 'https://picsum.photos/200/300',
-    shop: '아이디헤어문정점3',
-  },
-];
-
 const POST_LIMIT = 20;
 
 export default function PostsPage() {
@@ -88,7 +64,7 @@ export default function PostsPage() {
 
       <div className="flex flex-col gap-5 flex-1 min-h-0">
         <Tab options={POST_TABS} value={selectedTab} onChange={setSelectedTab} />
-        <TodayConsultantBanner banners={TODAY_CONSULTANT_BANNERS} />
+        <TodayConsultantBanner />
         <div className="flex-1 flex flex-col min-h-0 gap-2">
           <div className="flex-shrink-0">
             <ToggleChipGroup className="flex overflow-x-auto scrollbar-hide px-5">
